+++ tools/eslint-formatter-multi/src/index.ts
@@ -1,0 +1,3 @@
+export { default } from './lib/multiple-formats.js';
+export * from './lib/multiple-formats.js';
+export * from './lib/utils.js';